extracted quiz component and review cards
