+++ frontend/src/components/CodingChallengeCard.tsx
@@ -1,0 +1,13 @@
+import styles from '../styles/CodingChallengeCard.module.css'
+
+export default function ConceptReviewCard() {
+
+    // fetch concept quizzes
+        // need to split by open, locked, and completed
+    return (
+        <div className={styles.card}>
+            <h4>Title</h4>
+            <p>unit</p>
+        </div>
+    )
+}
