+++ frontend/src/pages/learn/quiz/[moduleId].tsx
@@ -2,10 +2,8 @@ import { useEffect, useState } from "react";
 import { useRouter } from "next/router";
 
 import QuizQuestion from "@/components/QuizQuestion";
-import ProgressBar from "@/components/ProgressBar";
-import styles from "@/styles/Quiz.module.css";
 import { getModuleTitle, getQuizQuestions, submitQuizScore } from "@/api/api";
-import QuizScore from "@/components/QuizScore";
+import Quiz from "@/components/Quiz";
 
 interface QuizQuestion {
     id: number;
@@ -17,26 +15,11 @@ interface QuizQuestion {
     }[];
 }
 
-const calculateProgressPercentage = (
-    numQuestions: number,
-    numCompleted: number
-) => {
-    return numQuestions > 0 ? (numCompleted / numQuestions) * 100 : 0;
-};
-
 const QuizPage = () => {
     const router = useRouter();
     const { moduleId } = router.query;
     const [questions, setQuestions] = useState<QuizQuestion[]>([]);
-    const [currentQuestionIndex, setCurrentQuestionIndex] = useState(0); // is the index reliable?
-    const [numCorrectAnswers, setNumCorrectAnswers] = useState<number>(0);
-    const [moduleTitle, setModuleTitle] = useState<string | null>(null);
-
-    const currentQuestion = questions[currentQuestionIndex];
-    const progressPercentage = calculateProgressPercentage(
-        questions.length,
-        currentQuestionIndex
-    );
+    const [moduleTitle, setModuleTitle] = useState<string>("");
 
     const fetchQuestions = async () => {
         try {
@@ -64,16 +47,7 @@ const QuizPage = () => {
         }
     }, [moduleId]);
 
-    // TODO: add feature to try missed questions again at the end of the quiz
-    const handleNextQuestion = (isCorrect: boolean) => {
-        // log correct answers and move to next question
-        if (isCorrect) {
-            setNumCorrectAnswers((prev) => prev + 1);
-        }
-        setCurrentQuestionIndex((prev) => prev + 1);
-    };
-
-    const handleContinue = async () => {
+    const handleContinue = async (numCorrectAnswers: number) => {
         const accuracy = (numCorrectAnswers / questions.length) * 100;
         try {
             await submitQuizScore(Number(moduleId), accuracy);
@@ -84,28 +58,12 @@ const QuizPage = () => {
     };
 
     return (
-        <div>
-            {currentQuestionIndex < questions.length ? (
-                <div className={styles.quizContainer}>
-                    <h1>{moduleTitle}</h1>
-                    <ProgressBar percentage={progressPercentage} />
-                    <QuizQuestion
-                        question={currentQuestion}
-                        handleNextQuestion={handleNextQuestion}
-                    />
-                </div>
-            ) : (
-                // If no questions left, show the score
-                <div className={styles.quizContainer}>
-                    <QuizScore
-                        numCorrectAnswers={numCorrectAnswers}
-                        numQuestions={questions.length}
-                        onContinue={handleContinue}
-                    />
-                </div>
-            )}
-        </div>
-    );
+        <Quiz
+            questions={questions}
+            moduleTitle={moduleTitle}
+            onSubmit={handleContinue}
+        />
+    )
 };
 
 export default QuizPage;
