+++ frontend/src/styles/CodingChallengeCard.module.css
@@ -1,0 +1,20 @@
+.card {
+    display: inline-flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    padding: 16px 20px;
+    gap: 1px;
+    margin-top: 9px;
+    margin-bottom: 14px;
+
+    border-radius: 12px;
+    border: 2px solid #d9d9d9;
+    background: #f7f7f7;
+}
+
+.card p {
+    color: #3C4DFF;
+    font-weight: 500; /* TODO: after loading fonts this should be 600 weight */
+    font-size: 12px;
+}
