+++ frontend/src/styles/WeeklyReviewCard.module.css
@@ -4,7 +4,7 @@
     justify-content: center;
     padding: 16px 20px;
     gap: 16px;
-    width: 232px;
+    width: 244px;
 
     border-radius: 12px;
     border: 2px solid #d9d9d9;
